Extract integration test configuration helper
