+++ tests/Plugin.Firebase.IntegrationTests/Storage/StorageFixture.cs
@@ -1,9 +1,7 @@
 using System.Net;
 using System.Text;
+using Plugin.Firebase.IntegrationTests;
 using Plugin.Firebase.Storage;
-#if ANDROID
-using AndroidRuntime = Android.Runtime;
-#endif
 
 namespace Plugin.Firebase.IntegrationTests.Storage
 {
@@ -122,76 +120,25 @@ namespace Plugin.Firebase.IntegrationTests.Storage
 
         private static bool UsesStorageEmulator()
         {
-            return string.Equals(
-                GetConfigurationValue("PLUGIN_FIREBASE_USE_STORAGE_EMULATOR", "debug.pluginfirebase.storage.use"),
-                "1",
-                StringComparison.Ordinal);
+            return IntegrationTestConfiguration.IsFeatureEnabled(
+                "PLUGIN_FIREBASE_USE_STORAGE_EMULATOR",
+                "debug.pluginfirebase.storage.use");
         }
 
         private static string GetStorageEmulatorHost()
         {
-            var host = GetConfigurationValue(
+            return IntegrationTestConfiguration.GetEmulatorHost(
                 "PLUGIN_FIREBASE_STORAGE_EMULATOR_HOST",
                 "debug.pluginfirebase.storage.host");
-            return string.IsNullOrWhiteSpace(host)
-                ? OperatingSystem.IsAndroid() ? "10.0.2.2" : "localhost"
-                : host;
         }
 
         private static int GetStorageEmulatorPort()
         {
-            var portValue = GetConfigurationValue(
+            return IntegrationTestConfiguration.GetEmulatorPort(
                 "PLUGIN_FIREBASE_STORAGE_EMULATOR_PORT",
-                "debug.pluginfirebase.storage.port");
-            return string.IsNullOrWhiteSpace(portValue) ? 9199 : int.Parse(portValue);
+                "debug.pluginfirebase.storage.port",
+                9199);
         }
-
-        private static string GetConfigurationValue(string environmentVariableName, string androidSystemPropertyName)
-        {
-            var environmentVariableValue = Environment.GetEnvironmentVariable(environmentVariableName);
-            if(!string.IsNullOrWhiteSpace(environmentVariableValue)) {
-                return environmentVariableValue;
-            }
-
-#if ANDROID
-            return GetAndroidSystemProperty(androidSystemPropertyName);
-#else
-            return null;
-#endif
-        }
-
-#if ANDROID
-        private static string GetAndroidSystemProperty(string propertyName)
-        {
-            IntPtr? propertyValuePointer = null;
-
-            try {
-                var systemPropertiesClass = AndroidRuntime.JNIEnv.FindClass("android/os/SystemProperties");
-                var getMethodId = AndroidRuntime.JNIEnv.GetStaticMethodID(
-                    systemPropertiesClass,
-                    "get",
-                    "(Ljava/lang/String;Ljava/lang/String;)Ljava/lang/String;");
-
-                using var propertyNameValue = new Java.Lang.String(propertyName);
-                using var defaultValue = new Java.Lang.String(string.Empty);
-                propertyValuePointer = AndroidRuntime.JNIEnv.CallStaticObjectMethod(
-                    systemPropertiesClass,
-                    getMethodId,
-                    new AndroidRuntime.JValue(propertyNameValue),
-                    new AndroidRuntime.JValue(defaultValue));
-
-                return AndroidRuntime.JNIEnv.GetString(
-                    propertyValuePointer.Value,
-                    AndroidRuntime.JniHandleOwnership.DoNotTransfer);
-            } catch {
-                return null;
-            } finally {
-                if(propertyValuePointer.HasValue && propertyValuePointer.Value != IntPtr.Zero) {
-                    AndroidRuntime.JNIEnv.DeleteLocalRef(propertyValuePointer.Value);
-                }
-            }
-        }
-#endif
 
         [Fact]
         public async Task uploads_via_byte_array()
